Create final project
Final project for class. The instructions were to create a fake business which has a problem needing to be solved. My fake business was that of Ralph's Computer Assembly, a hard working computer assembler named Ralph has a steadily growing clientele but an antiquated data system. The above code will help Ralph to create an online database with relationships and functions to keep his info up to date. 

Admittedly, there are MANY better ways than this for coding a database. For this being my first undertaking of a project with this many variables within SQL, I was just happy to get a functioning set of queries. 

The drop table function at the beginning was a requirement in order to create a repeatable script, this code can be copied and pasted in whole and will produce the same results each time.